# backend/storage/artifacts/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Ecosystem Multi-Service Core API",
    "description": "## ระบบบริการส่วนหลัง AI Ecosystem (Backend Multi-Service Architecture)\n\nแอปพลิเคชันนี้ทำหน้าที่เป็น **Core Backend Server** สำหรับเชื่อมต่อและให้บริการผ่านองค์ประกอบต่างๆ:\n\n* **Relational Database**: PostgreSQL สำหรับเก็บข้อมูลบัญชีผู้ใช้และ Metadata\n* **Cloud Object Storage**: MinIO (S3 Compatible) สำหรับเก็บไฟล์สื่อและรูปภาพ\n* **In-Memory Cache & Message Broker**: Redis & ARQ สำหรับระบบแคชและคิวงานเบื้องหลัง\n* **Data Annotation Platform**: Label Studio สำหรับจัดการชุดข้อมูลและติดฉลากสำหรับ AI/ML\n\n--- \n### 🔐 การยืนยันตัวตน (Authentication)\nปุ่ม **Authorize** ด้านบนสุดใช้ใส่ค่า **Bearer Access Token** ที่ได้จากการเรียก API `/auth/login`",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "AI Ecosystem Developer Team",
      "url": "https://example.com/support",
      "email": "dev-team@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/storage/objects": {
      "get": {
        "tags": [
          "MinIO Object Storage"
        ],
        "summary": "เรียกดูรายการไฟล์ใน Bucket (List Objects)",
        "description": "เรียกดูรายชื่อวัตถุ/ไฟล์ทั้งหมดใน Bucket ที่ระบุ",
        "operationId": "list_bucket_objects_storage_objects_get",
        "parameters": [
          {
            "name": "bucket_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ชื่อ Bucket ที่ต้องการค้นหา",
              "title": "Bucket Name"
            },
            "description": "ชื่อ Bucket ที่ต้องการค้นหา"
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Prefix เส้นทางโฟลเดอร์ (ถ้ามี)",
              "title": "Prefix"
            },
            "description": "Prefix เส้นทางโฟลเดอร์ (ถ้ามี)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObjectItem"
                  },
                  "title": "Response List Bucket Objects Storage Objects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/storage/presigned-url": {
      "post": {
        "tags": [
          "MinIO Object Storage"
        ],
        "summary": "สร้าง Presigned URL สำหรับดาวน์โหลดไฟล์แบบจำกัดเวลา",
        "description": "สร้างลิงก์ดาวน์โหลดไฟล์ชั่วคราวผ่าน MinIO Presigned GET Object API",
        "operationId": "generate_object_presigned_url_storage_presigned_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/cache": {
      "post": {
        "tags": [
          "Redis Cache & ARQ Tasks"
        ],
        "summary": "บันทึกข้อมูลเข้า Redis Cache (Set Key-Value with TTL)",
        "description": "บันทึกข้อมูล Key-Value ลงใน Redis Cache พร้อมกำหนดเวลาหมดอายุ (TTL)",
        "operationId": "set_redis_cache_tasks_cache_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/cache/{key}": {
      "get": {
        "tags": [
          "Redis Cache & ARQ Tasks"
        ],
        "summary": "ดึงข้อมูลจาก Redis Cache (Get Key-Value)",
        "description": "อ่านข้อมูลจาก Redis Cache ตาม Key ที่ระบุ",
        "operationId": "get_redis_cache_tasks_cache__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/enqueue": {
      "post": {
        "tags": [
          "Redis Cache & ARQ Tasks"
        ],
        "summary": "ส่งงานเข้าคิวประมวลผลเบื้องหลัง (Enqueue ARQ Task)",
        "description": "ส่งคำสั่งเข้าคิวงานเบื้องหลัง (ARQ Task Queue) เพื่อให้ Background Worker ดึงไปประมวลผลแบบ Asynchronous",
        "operationId": "enqueue_background_task_tasks_enqueue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CacheResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "is_cached": {
            "type": "boolean",
            "title": "Is Cached"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "is_cached"
        ],
        "title": "CacheResponse"
      },
      "CacheSetRequest": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "example": "system:config:theme"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "example": "dark"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds",
            "description": "เวลาหมดอายุของข้อมูล (วินาที)",
            "default": 300
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "CacheSetRequest"
      },
...
      "EnqueueTaskRequest": {
        "properties": {
          "job_name": {
            "type": "string",
            "title": "Job Name",
            "default": "simple_work",
            "example": "simple_work"
          },
          "job_data": {
            "type": "string",
            "title": "Job Data",
            "example": "Processing Dataset Batch #102"
          }
        },
        "type": "object",
        "required": [
          "job_data"
        ],
        "title": "EnqueueTaskRequest"
      },
      "EnqueueTaskResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "message"
        ],
        "title": "EnqueueTaskResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObjectItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "avatars/user-123.jpg"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "example": 102400
          },
          "last_modified": {
            "type": "string",
            "title": "Last Modified"
          }
        },
        "type": "object",
        "required": [
          "name",
          "size",
          "last_modified"
        ],
        "title": "ObjectItem"
      },
      "PresignedUrlRequest": {
        "properties": {
          "bucket_name": {
            "type": "string",
            "title": "Bucket Name",
            "example": "profile-images"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name",
            "example": "avatars/user-123.jpg"
          },
          "expires_hours": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Expires Hours",
            "description": "จำนวนชั่วโมงที่ URL สามารถใช้งานได้",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "bucket_name",
          "object_name"
        ],
        "title": "PresignedUrlRequest"
      },
      "PresignedUrlResponse": {
        "properties": {
          "bucket_name": {
            "type": "string",
            "title": "Bucket Name"
          },
          "object_name": {
            "type": "string",
            "title": "Object Name"
          },
          "presigned_url": {
            "type": "string",
            "title": "Presigned Url"
          },
          "expires_in_hours": {
            "type": "integer",
            "title": "Expires In Hours"
          }
        },
        "type": "object",
        "required": [
          "bucket_name",
          "object_name",
          "presigned_url",
          "expires_in_hours"
        ],
        "title": "PresignedUrlResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}